decomp: repair three uncompilable wip seeds (08069B78, 0806C78C, 080411EC)
- sub_08069B78: u8 args -> u32 (header signature)
- sub_0806C78C: header return void -> s32, first arg kept as void*
- sub_080411EC: _08073C48 needs a third (fn) argument
All three now score (11-17%) instead of failing to build.

+++ src/wip/sub_08069B78.c
@@ -1,7 +1,7 @@
 #include "global.h"
 
 // @ 0x08069b78
-void sub_08069B78(u8 a, u8 b, u8 c, u8 d)
+void sub_08069B78(u32 a, u32 b, u32 c, u32 d)
 {
     u8 *p;
